refactor: move details and accordion src CSS to styles folders
