+++ packages/accordion/src/styles/vaadin-accordion-heading-base-styles.d.ts
