+++ packages/accordion/src/styles/vaadin-accordion-heading-core-styles.js
