+++ packages/accordion/src/styles/vaadin-accordion-panel-base-styles.d.ts
