+++ packages/accordion/src/styles/vaadin-accordion-panel-core-styles.js
