+++ packages/accordion/src/vaadin-accordion-heading.js
@@ -10,7 +10,7 @@ import { DirMixin } from '@vaadin/component-base/src/dir-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { accordionHeading } from './vaadin-accordion-heading-core-styles.js';
+import { accordionHeading } from './styles/vaadin-accordion-heading-core-styles.js';
 
 /**
  * The accordion heading element.
