+++ packages/details/src/styles/vaadin-details-summary-base-styles.js
