+++ packages/details/src/styles/vaadin-details-summary-core-styles.js
